Refactor API methods for consistency and add CSRF protection for JWT cookies; remove obsolete test files and enhance user form tests
